requirements.txt
Django==1.8.4
Pillow==2.9.0
psycopg2==2.6.1
djangorestframework==3.2.3
